Make the outline editor actually feel like Notes: Enter, Backspace, fast
Enter now splits the line at the cursor and creates a real new sibling
(matching bold spans on both halves via range-based DOM splitting, not
just a naive text split). Backspace at the start of a line merges it
into the previous one and removes it, mirroring how Notes behaves.

Discovered and fixed two real problems along the way, not just the
missing keybindings:

- Every Supabase round-trip here costs ~1s, and the original
  implementation chained 3-4 of them sequentially per Enter (renumber
  every following sibling, save, insert, full reload) - 5 full seconds
  before you could type the next line. Fixed by switching todo_outline.
  position from int to numeric (insert at a fractional position instead
  of shifting every sibling after it) and applying optimistic local
  updates instead of waiting on a round-trip to re-render - now ~15ms.

- Optimistic inserts get a temp- id so they're immediately editable;
  once the real uuid comes back it's relabeled in place (not a
  re-render), so an in-progress edit is never disturbed. A second Enter
  fired before the first insert confirms now awaits that pending id
  instead of sending "temp-xxx" to Postgres as a uuid.

Tested rapid multi-line typing (the scenario that exposed both bugs),
single-line split with bold preservation, and backspace-merge; confirmed
Sanjay's tier view and the CRM are unaffected.

+++ schema.sql
@@ -68,12 +68,15 @@ create table if not exists public.todo_items (
 -- how THIS node renders inside its parent's list: 'none' (a heading/paragraph, no marker),
 -- 'numbered', or 'dashed'. content may contain **bold** spans, rendered as <strong>. Only
 -- list_style != 'none' nodes are checkable (done); checking one collapses it into a
--- "Completed" section in the UI without changing its place in the tree.
+-- "Completed" section in the UI without changing its place in the tree. position is numeric
+-- (not int) so the app can insert a new line between two existing siblings (e.g. 5.5 between
+-- 5 and 6) without renumbering every row after it — that's what makes pressing Enter mid-list
+-- fast instead of cascading position updates across the whole section.
 create table if not exists public.todo_outline (
   id uuid primary key default gen_random_uuid(),
   owner_id uuid not null references auth.users(id) default auth.uid(),
   parent_id uuid references public.todo_outline(id) on delete cascade,
-  position int not null default 0,
+  position numeric not null default 0,
   list_style text not null default 'none' check (list_style in ('none', 'numbered', 'dashed')),
   content text not null default '',
   done boolean not null default false,
